LayoutTests/inspector/extensions/extensions-settings-expected.txt

Tests WebInspector extension API

Started extension.
Running tests...
RUNNING TEST: extension_testCreateCheckbox
Checkbox created
Checkbox was checked
All tests done.
